Backend clean up and adding user data db
