+++ backend/main.py
@@ -1,35 +1,22 @@
-from flask import Flask, render_template, redirect, url_for, request, flash, jsonify
+from flask import request, jsonify
 from werkzeug.security import generate_password_hash, check_password_hash
-from flask_sqlalchemy import SQLAlchemy
-from flask_login import LoginManager, login_user, logout_user, login_required, UserMixin
-from flask_cors import CORS, cross_origin
+from flask_login import LoginManager, login_user
+from flask_cors import cross_origin
+from . import app
+from . import db
+from .models import User, UserData
 
-app = Flask(__name__)
-app.config.from_object(__name__)
-
-CORS(app, resources={r"/api/*": {"origins": "*"}})# CORS allowed for all domains on all routes
-
-app.secret_key = b'_5#y2L"F4Q8z\n\xec]/'
-app.config['SQLALCHEMY_DATABASE_URI'] = 'sqlite:///db.sqlite'
-
-db = SQLAlchemy(app)
-
-class User(db.Model, UserMixin):
-    id = db.Column(db.Integer, primary_key=True) # primary keys are required by SQLAlchemy
-    password = db.Column(db.String(100))
-    name = db.Column(db.String(1000), unique=True)
-
-with app.app_context():
-    db.create_all()
 
 login_manager = LoginManager()
 login_manager.login_view = 'login'
 login_manager.init_app(app)
 
+
 @login_manager.user_loader
 def load_user(user_id):
     # since the user_id is just the primary key of our user table, use it in the query for the user
     return User.query.get(int(user_id))
+
 
 @app.route("/api/login", methods=['POST', 'GET'])
 @cross_origin()
@@ -49,14 +36,14 @@ def login():
 
         else:
             login_user(user)
+            user_data=UserData.query.get(user.id)
+            response_object['userId']=user.id
+            response_object['userName']=user.name
+            response_object['userDroplets']=user_data.droplets
+            response_object['userStreak']=user_data.streak
             return jsonify(response_object)
 
-"""
-@app.route("/home")
-@login_required
-def home():
-    return 'you are in the home page'
-"""
+
 @app.route('/api/signup', methods=['POST', 'GET'])
 @cross_origin()
 def signup():
@@ -72,28 +59,15 @@ def signup():
 
         # create new user with the form data. Hash the password so plaintext version isn't saved.
         new_user = User(name=name, password=generate_password_hash(password, method='sha256'))
+        
 
         # add the new user to the database
         db.session.add(new_user)
         db.session.commit()
 
         user = User.query.filter_by(name=name).first()
-        login_user(user)
+        user_data=UserData(id=user.id, droplets=0, streak=0, challenge=0)
+        db.session.add(user_data)
+        db.session.commit()
 
     return response_object
-"""
-@app.route('/modes/<user_id>')
-@login_required
-def modes():
-    modes={'car', 'bus'}
-    return render_template('modes.html', data=modes)
-
-
-@app.route('/logout')
-@login_required
-def logout():
-    logout_user()
-    return redirect(url_for('login'))
-"""
-if __name__=="__main__":
-    app.run(debug=True)
